Add a URL normalisation step
Some websites have multiple URLs for the same thing, which can be
normalised.  This can't be done in the es_presenter, as we want the
scraper to see the normalised URL to prevent redundant pattern cases.

Reindexing will renormalise the URLs.  Importing will not, because
normalised URLs may trigger different scraping behaviour, and I don't
want to do scraping on index restore.